salary of a government employee(OOP)
This program can calculate and display the salary of a government employee. This program first takes the personal information of the person and calculates the salary of that person based on the points of the job class and the person's rank and prints the salary slip of the person. I have used the rights of the government departments of my country as well as advanced programming methods called object orientation in writing this program. It still needs further development and adding more parts